#11 Figured out that stubbing  causes a bad access on 64 bit achitecture because of the assignment of the given array literal.
I'd assumed when calling andReturn, the value passed was retained / copied / managed by the framework but it seems that it isn't reliable held in memory. There is no bad access for example, if I declare the array as a variable like so:

    NSArray* recognizers = @[panGestureRecognizer]
    OCMStub([superview gestureRecognizers]).andReturn(recognizers);

because its autoreleased at the end of the block. For now I've opted just to add gesture recognizers via the defined interface in UIView - seems much simpler