#11 Added fixture for implementation of the I3Collection
According the OCMock docs, protocol mocks work in the same way as class mocks and therefore have the same limitations when it comes to mocking the same protcol multiple times at once. We've worked around that my partially mocking an instance of a class that implements that protocol. I'm not 100% sure whether partially mocking 2 instances of the same class as once has the same limitations (the docs seems a bit ambiguous).

Also removed unnecessary property in the I3TableView